User menu

+++ MVVM/ViewModel/TourViewModel.cs
@@ -2,14 +2,8 @@
 using ExploreRussia.MVVM.Model;
 using ExploreRussia.MVVM.Repositories;
 using ExploreRussia.MVVM.View;
-using System.Collections;
 using System.Collections.Generic;
-using System.Collections.ObjectModel;
-using System.Net.Http.Headers;
-using System.Runtime.Serialization;
 using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Input;
 
 namespace ExploreRussia.MVVM.ViewModel
 {
@@ -23,18 +17,18 @@ namespace ExploreRussia.MVVM.ViewModel
 
         //команды
         public RelayCommand TopViewEditCommand { get; set; }
-        public RelayCommand TopViewAddCommand {  get; set; }
+        public RelayCommand TopViewAddCommand { get; set; }
         public RelayCommand DeleteTourCommand { get; set; }
 
 
-        public object SelectedItem 
-        { 
+        public object SelectedItem
+        {
             get { return _selectedItem; }
             set
             {
                 _selectedItem = (TourModel)value;
                 OnPropertyChanged(nameof(SelectedItem));
-            } 
+            }
         }
 
         public object TopView
@@ -62,8 +56,8 @@ namespace ExploreRussia.MVVM.ViewModel
         {
             TourRepository tours = new TourRepository();
             ItemsSource = tours.GetAll();
-            TopViewEditCommand = new RelayCommand(o => 
-            { 
+            TopViewEditCommand = new RelayCommand(o =>
+            {
                 if (SelectedItem != null)
                 {
                     AddEditWindow addEditWindow = new AddEditWindow((TourModel)SelectedItem);
@@ -83,12 +77,15 @@ namespace ExploreRussia.MVVM.ViewModel
             {
                 if (SelectedItem != null)
                 {
-                    TourRepository tourRepository = new TourRepository();
-                    tourRepository.Remove((TourModel)SelectedItem); 
+                    if (MessageBox.Show("Вы уверены, что хотите удалить эту запись?", "Удаление", MessageBoxButton.YesNo) == MessageBoxResult.Yes)
+                    {
+                        TourRepository tourRepository = new TourRepository();
+                        tourRepository.Remove((TourModel)SelectedItem);
+                    }
                     ItemsSource = tours.GetAll();
                 }
             });
-            
+
         }
 
     }
